+++ Sources/LoadableViews/LoadableView+macOS.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-#if canImport(AppKit)
+#if !targetEnvironment(macCatalyst) && canImport(AppKit)
 import AppKit
 
 extension NibLoadableProtocol {